requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argostranslate==1.7.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backoff==2.1.2
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
cachetools==4.2.4
cbor==1.0.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
chest==0.2.3
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.5
ctranslate2==2.17.0
cwl-eval==1.0.10
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.5.1
docker-pycreds==0.4.0
emoji==2.2.0
entrypoints==0.4
executing==1.0.0
fairseq==0.12.1
fastBPE==0.1.0
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.2
frozenlist==1.3.1
gcld3==3.0.13
gitdb==4.0.9
GitPython==3.1.27
google-api-core==1.33.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
google-cloud-vision==3.1.2
googleapis-common-protos==1.56.4
gql==3.4.0
graphql-core==3.2.1
grpcio==1.49.0
grpcio-status==1.48.1
HeapDict==1.0.1
huggingface-hub==0.9.1
hydra-core==1.0.7
idna==3.4
ijson==3.1.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.3
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
ir-datasets==0.5.3
ir-measures==0.3.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-http-over-ws==0.0.8
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
kiwisolver==1.4.4
lxml==4.9.1
lz4==4.0.2
MarkupSafe==2.1.1
matchpy==0.5.5
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==8.14.0
multidict==6.0.2
multiset==2.1.1
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
nptyping==1.4.4
numpy==1.23.3
omegaconf==2.0.6
onnxruntime==1.14.1
onnx==1.13.1
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.3
portalocker==2.6.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.20.2
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyautocorpus==0.1.8
pybind11==2.10.1
pycparser==2.21
Pygments==2.13.0
pyjnius==1.4.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-terrier==0.9.1
pytrec-eval-terrier==0.5.4
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
regex==2022.9.13
requests==2.28.1
rich==12.5.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
statsmodels==0.13.2
tabulate==0.9.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.0
transliterate==1.10.2
trec-car-tools==2.6
typing_extensions==4.3.0
typish==1.9.3
Unidecode==1.3.6
unlzw3==0.2.1
urllib3==1.26.12
wandb==0.13.3
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
wget==3.2
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
zlib-state==0.1.5
